Create AndroidKinClient to conform to previous Context aware interface of KinClient. Also refactored KinClient -> KinClientImpl and created interface.
