+++ sample/src/main/java/kin/sdk/sample/KinClientSampleApplication.java
@@ -4,12 +4,9 @@ import android.app.Application;
 import android.os.StrictMode;
 import android.os.StrictMode.VmPolicy;
 
-import kin.sdk.AndroidMainHandler;
+import kin.sdk.AndroidKinClient;
 import kin.sdk.Environment;
 import kin.sdk.KinClient;
-import kin.sdk.KinStorageFactory;
-import kin.sdk.KinStorageFactoryImpl;
-import kin.utils.MainHandler;
 
 public class KinClientSampleApplication extends Application {
 
@@ -20,12 +17,8 @@ public class KinClientSampleApplication extends Application {
 
     private KinClient kinClient = null;
 
-    private KinStorageFactory kinStorageFactory = new KinStorageFactoryImpl(this);
-    private MainHandler androidMainHandler = new AndroidMainHandler();
-
     public KinClient createKinClient(NetWorkType type, String appId) {
-        kinClient = new KinClient(kinStorageFactory,
-                androidMainHandler,
+        kinClient = new AndroidKinClient(this,
                 type == NetWorkType.MAIN ? Environment.PRODUCTION : Environment.TEST,
                 appId,
                 "sample_app");
